fix cmd

+++ plexhelp-commander-cli/src/main/java/net/cserny/task/clean/CleanEmptyService.java
@@ -51,7 +51,7 @@ public class CleanEmptyService {
                 }
             }
 
-            if (!containsMediaFiles) {
+            if (!containsMediaFiles && !mediaFolder.path().toString().equals(path)) {
                 log.info("No files found in media folder {}, deleting...", mediaFolder);
                 localPathHandler.deleteDirectory(mediaFolder);
             }
